requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
openai>=1.30.0
PyMuPDF==1.23.8
python-docx==1.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
